+++ testutil/testutil.go
@@ -16,7 +16,6 @@ package testutil
 
 import (
 	"bytes"
-	"encoding/json"
 	"io"
 	"io/ioutil"
 	"net"
@@ -33,22 +32,6 @@ import (
 type portpair struct {
 	first string
 	last  string
-}
-
-// GenerateNormalizedJSON generates a normalized JSON from the string
-// given to the function. Useful to compare JSON contents that
-// may have differences due to formatting. It returns nil in case of
-// invalid JSON.
-func GenerateNormalizedJSON(t *testing.T, jsonStr string) string {
-	var i interface{}
-
-	err := json.Unmarshal([]byte(jsonStr), &i)
-	require.NoError(t, err)
-
-	n, err := json.Marshal(i)
-	require.NoError(t, err)
-
-	return string(n)
 }
 
 // TempSocketName provides a temporary Unix socket name for testing.